+++ CHANGELOG.md
@@ -40,8 +40,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - [#53](https://github.com/alleslabs/celatone-frontend/pull/53) Show contract description in contract details page
-
-- [#46](https://github.com/alleslabs/celatone-frontend/pull/46) Wireup instantiate info
+- [#58](https://github.com/alleslabs/celatone-frontend/pull/58) Wireup top section in contract details page 
+- [#54](https://github.com/alleslabs/celatone-frontend/pull/54) Render execute cmds shortcut in contract details page
+- [#46](https://github.com/alleslabs/celatone-frontend/pull/46) Wireup instantiate info in contract details page
 - [#55](https://github.com/alleslabs/celatone-frontend/pull/55) Add "Add To List / Edit" button to edit offchain details on query and execute pages
 - [#44](https://github.com/alleslabs/celatone-frontend/pull/44) Render query cmds shortcut in contract details page
 - [#38](https://github.com/alleslabs/celatone-frontend/pull/38) Show execute msg cmds when wallet is not connected